added restricted page functionality

+++ studybud/base/views.py
@@ -2,7 +2,9 @@ from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.http import HttpResponse
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -10,6 +12,9 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        redirect("home")
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("pass")
@@ -58,6 +63,7 @@ def room(request, id):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="/login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -70,9 +76,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def updateRoom(request, id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -83,8 +93,13 @@ def updateRoom(request, id):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def deleteRoom(request, id):
     room = Room.objects.get(id=id)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
